fixed few bugs
&& operator not always work as I intended, instead made some if's

+++ controllers/usersController.js
@@ -2,23 +2,28 @@ const User = require("../model/User")
 
 const getAllUsers = async (req, res) => {
   const users = await User.find()
-  !users && res.status(204).json({ message: "No users found" })
+  if (!users) return res.status(204).json({ message: "No users found" })
   res.json(users)
 }
 
 const deleteUser = async (req, res) => {
-  !req?.body?.id && res.status(400).json({ message: "User ID required" })
+  if (!req?.body?.id)
+    return res.status(400).json({ message: "User ID required" })
   const user = await User.findOne({ _id: req.body.id }).exec()
-  !user && res.status(204).json({ message: `User ID ${req.body.id} not found` })
+  if (!user)
+    return res.status(204).json({ message: `User ID ${req.body.id} not found` })
   const result = await user.deleteOne({ _id: req.body.id })
   res.json(result)
 }
 
 const getUser = async (req, res) => {
-  !req?.params?.id && res.status(400).json({ message: "User ID required" })
+  if (!req?.params?.id)
+    return res.status(400).json({ message: "User ID required" })
   const user = await User.findOne({ _id: req.params.id }).exec()
-  !user &&
-    res.status(204).json({ message: `User ID ${req.params.id} not found` })
+  if (!user)
+    return res
+      .status(204)
+      .json({ message: `User ID ${req.params.id} not found` })
   res.json(user)
 }
